Basic styles elf

+++ components/RecipeDetails/RecipeDetails.js
@@ -1,13 +1,16 @@
+/* eslint-disable import/no-extraneous-dependencies */
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { Button, Card } from 'react-bootstrap';
+import { Card } from 'react-bootstrap';
+import { Heart, HeartFill } from 'react-bootstrap-icons';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import {
   deleteRecipe, favoriteRecipe, getSingleRecipe, unfavoriteRecipe,
 } from '../../utils/data/recipeData';
 import { useAuth } from '../../utils/context/authContext';
+import styles from './RecipeDetails.module.css';
 
 export default function RecipeDetails({ id }) {
   const [recipe, setRecipe] = useState({});
@@ -23,7 +26,9 @@ export default function RecipeDetails({ id }) {
   }, [id, user]);
 
   const handleDelete = () => {
-    deleteRecipe(id).then(() => router.push('/'));
+    if (window.confirm(`Are you sure you want to delete "${recipe.title}"? This action cannot be undone.`)) {
+      deleteRecipe(id).then(() => router.push('/'));
+    }
   };
 
   const handleFavorite = () => {
@@ -36,25 +41,44 @@ export default function RecipeDetails({ id }) {
   };
 
   return (
-    <>
-      <Card style={{ width: '18rem' }}>
+    <div className={styles.recipeDetailsWrapper}>
+      <Card className={styles.card}>
         <Card.Body>
-          <Button variant="secondary" size="sm" onClick={handleFavorite}>
-            {recipe.is_favorite ? 'Unfavorite' : 'Favorite'}
-          </Button>
-          <Card.Title>{recipe.title}</Card.Title>
+          <section className={styles.favoriteDiv}>
+            <Card.Title className={styles.title}>{recipe.title}</Card.Title>
+            <button type="button" className={styles.favoriteBtn} onClick={handleFavorite}>
+              {recipe.is_favorite
+                ? <HeartFill className={styles.heartFill} />
+                : <Heart className={styles.heart} />}
+              Favorite
+            </button>
+          </section>
+          <hr />
+          <Card.Title>Ingredients:</Card.Title>
           {recipe.recipe_oils?.map((oil) => (
-            <Card.Subtitle key={oil.id} className="mb-2 text-muted">{oil.oil_name}: {oil.amount} oz</Card.Subtitle>
+            <section className={styles.ingredientsSection}>
+              <Card.Text key={oil.id} className={styles.ingredients}>{oil.oil_name}:</Card.Text>
+              <Card.Text key={oil.id} className={styles.ingredients}>{Number(oil.amount)} oz</Card.Text>
+            </section>
           ))}
-          <Card.Subtitle className="mb-2 text-muted">Water: {recipe.water_amount} oz</Card.Subtitle>
-          <Card.Subtitle className="mb-2 text-muted">Lye: {recipe.lye_amount} oz</Card.Subtitle>
-          <Card.Subtitle className="mb-2 text-muted">Super fat: {recipe.super_fat * 100}%</Card.Subtitle>
-          <Card.Text>Description: {recipe.description}</Card.Text>
-          <Card.Text>Notes: {recipe.notes}</Card.Text>
-          <Card.Text>{recipe.public ? 'Public' : 'Private'}</Card.Text>
+          <section className={styles.ingredientsSection}>
+            <Card.Text className={styles.ingredients}>Water:</Card.Text>
+            <Card.Text className={styles.ingredients}>{Number(recipe.water_amount)} oz</Card.Text>
+          </section>
+          <section className={styles.ingredientsSection}>
+            <Card.Text className={styles.ingredients}>Lye (NaOH):</Card.Text>
+            <Card.Text className={styles.ingredients}>{Number(recipe.lye_amount)} oz</Card.Text>
+          </section>
+          <hr />
+          <Card.Title>Description:</Card.Title>
+          <Card.Text>{recipe.description}</Card.Text>
+          <hr />
+          <Card.Title>Notes:</Card.Title>
+          <Card.Text>{recipe.notes}</Card.Text>
           {user.uid === recipe.maker?.uid
             ? (
               <>
+                <hr />
                 <Link href={`/recipe/edit/${id}`} passHref>
                   <Card.Link href="#">Edit</Card.Link>
                 </Link><Card.Link href="#" onClick={handleDelete}>Delete</Card.Link>
@@ -63,7 +87,7 @@ export default function RecipeDetails({ id }) {
             : ''}
         </Card.Body>
       </Card>
-    </>
+    </div>
   );
 }
